Add files via upload
Fixed bug for AI meshes that would interfere with multiple materials being assigned to a mesh.
Fixed bug with ILLUM being allowed in materials in the .e file, which causes an error with Meshbld which doesn't support that. Future release will see MeshUp being used to convert meshes to "version 2" and will see if MatTweak can be called and automated to re-add illumination.