Add files via upload
Moved most of the object export params to the 'Scene' panel. This means they will be saved with each Blender file (a.k.a scene) and applied to any mesh(es) that you export. This will help with subsequent exporting for objects which don't use the default values. 
Note that is is possible to have these params linked to each mesh (i.e. allowing different values for each one), but it would introduce too much complexity, with the addon not knowing which set of params to apply to a scene with mulitple meshes to be exported as one object (e.g. jointed objects, or static objects whose parts are kept separate just in Blender).

Documented a button in the above panel which allows users to open the Config file from within Blender (program must be restarted for changes to take effect).

Fixed an error with using Coplanar Limit - the exporter was actually reading the coplanar number for it's 'use coplanar limit' input, instead 'true or false' value of the checkbox. The code worked without error just because Python treats 1.0 (the default value) the the same as true (and 0 is the same as false).