Add files via upload
Updated the Exporter - it had been using the coplanar value instead of the 'use coplanar' flag. Python treads values of 1 and 0 the same as true and false. 1 is the default, so very few people will have had any problems.
Updated the Importer - has an option to apply a workaround for old graphics cards which cannot properly show textures in version 4.2.